# Conflicts:
#	Assets/Prefabs/Skill/Attack/FanShapeFireSkill.prefab
#	Assets/Prefabs/Skill/Attack/FanShapeFireSkill.prefab.meta
#	Assets/Prefabs/Skill/Attack/FlameThrower.prefab
#	Assets/Prefabs/Skill/Attack/FlameThrower.prefab.meta
#	Assets/Prefabs/Skill/mon/FanShapeFireSkill.prefab
#	Assets/Prefabs/Skill/mon/FanShapeFireSkill.prefab.meta
#	Assets/Prefabs/Skill/mon/FlameThrower.prefab
#	Assets/Prefabs/Skill/mon/FlameThrower.prefab.meta
#	Assets/Resources/PerformanceTestRunInfo.json
#	Assets/Scripts/GAS/SkILL/FanShapeFireSkill.prefab
#	Assets/Scripts/GAS/SkILL/FanShapeFireSkill.prefab.meta
#	Assets/Scripts/GAS/SkILL/FlameThrower.prefab
#	Assets/Scripts/GAS/SkILL/FlameThrower.prefab.meta

+++ Assets/Scripts/Ui/ItemdescriptionView.cs
@@ -37,7 +37,7 @@ public class ItemdescriptionView : MonoBehaviour
 
     }
 
-    private void EnableCheck(LocalizedText tmp, int value)
+    private void EnableCheck(LocalizedText tmp, float value)
     {
         tmp.gameObject.SetActive(value > 0);
         tmp.UpdateLocalizedText(value);
